# Optimize StopWordFilter 