--- a/src/Transformers/StopWordFilter.php
+++ b/src/Transformers/StopWordFilter.php
@@ -24,18 +24,23 @@ class StopWordFilter extends RegexFilter
      */
     public function __construct(array $stopWords = [])
     {
-        $patterns = [];
-
-        foreach ($stopWords as $word) {
-            if (!is_string($word)) {
+        foreach ($stopWords as &$word) {
+            if (!is_string($word) || empty($word)) {
                 throw new InvalidArgumentException('Stop word must be a'
-                    . ' string, ' . gettype($word) . ' found.');
+                    . 'non-empty string, ' . gettype($word) . ' found.');
             }
 
-            $patterns[] = '/\b' . preg_quote($word, '/') . '\b/u';
+            $word = preg_quote($word, '/');
         }
 
-        parent::__construct($patterns);
+        if (empty($stopWords)) {
+            throw new InvalidArgumentException('Must specify at least'
+                . ' 1 stop word');
+        }
+
+        $pattern = sprintf('/\b(%s)\b/u', implode('|', $stopWords));
+
+        parent::__construct([$pattern]);
     }
 
     /**