#### SUMMARY:

- Optimize the `StopWordFilter` implementation and provide a benchmark script. 

- <s>Add the default stopwords that lucene uses: [`apache-lucene`::`getDefaultStopSet()`](https://github.com/apache/lucene-solr/blob/master/lucene/analysis/common/src/java/org/apache/lucene/analysis/en/EnglishAnalyzer.java#L46...L55)</s>



---

#### BENCHMARK:

The existing implementation completes in ~9 seconds, whilst the new implementation completes in ~1.3 seconds. The gain is achieved by using a single pattern to search for all the provided stopwords instead using separate patterns and passes to search for each individual stopword.

######  📉    EXISITING IMPLEMENTATION:
```
$ ./vendor/bin/phpbench run benchmarks/Transformers/StopWordFilterNormalizerBench.php --report=default
PhpBench. Running benchmarks.
Using configuration file: /Users/chris/Developer/github.com/simplechris/RubixML/phpbench.json.dist

\Rubix\ML\Benchmarks\Transformers\StopWordFilterNormalizerBench (#0 apply)

#0  9,841.620 9,757.097 9,627.247 (ms) [μ Mo]/r: 9,741.988 9,771.980 μRSD/r: 0.91%

1 subjects, 3 iterations, 1 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 9.627 [9.742 9.772] 9.842 (s)
⅀T: 29.226s μSD/r 0.088s μRSD/r: 0.905%
suite: 1343ddb0d29ec507359726bd15b23765d8438452, date: 2020-09-23, stime: 12:18:34
+-------------------------------+---------+-----+------+------+--------------+-------------+--------------+----------------+
| benchmark                     | subject | set | revs | iter | mem_peak     | time_rev    | comp_z_value | comp_deviation |
+-------------------------------+---------+-----+------+------+--------------+-------------+--------------+----------------+
| StopWordFilterNormalizerBench | apply   | 0   | 1    | 0    | 265,801,648b | 9,841.620ms | +1.13σ       | +1.02%         |
| StopWordFilterNormalizerBench | apply   | 0   | 1    | 1    | 265,801,648b | 9,757.097ms | +0.17σ       | +0.16%         |
| StopWordFilterNormalizerBench | apply   | 0   | 1    | 2    | 265,801,648b | 9,627.247ms | -1.3σ        | -1.18%         |
+-------------------------------+---------+-----+------+------+--------------+-------------+--------------+----------------+
```

⠀
######  📈    NEW IMPLEMENTATION:

```
$ ./vendor/bin/phpbench run benchmarks/Transformers/StopWordFilterNormalizerBench.php --report=default
PhpBench. Running benchmarks.
Using configuration file: /Users/chris/Developer/github.com/simplechris/RubixML/phpbench.json.dist

\Rubix\ML\Benchmarks\Transformers\StopWordFilterNormalizerBench (#0 apply)

#0  1,285.937 1,250.089 1,284.863 (ms) [μ Mo]/r: 1,273.630 1,283.271 μRSD/r: 1.31%

1 subjects, 3 iterations, 1 revs, 0 rejects, 0 failures, 0 warnings
(best [mean mode] worst) = 1.250 [1.274 1.283] 1.286 (s)
⅀T: 3.821s μSD/r 0.017s μRSD/r: 1.307%
suite: 1343ddb23cb4e6934212a6d152e82f3149636f2f, date: 2020-09-23, stime: 12:20:14
+-------------------------------+---------+-----+------+------+--------------+-------------+--------------+----------------+
| benchmark                     | subject | set | revs | iter | mem_peak     | time_rev    | comp_z_value | comp_deviation |
+-------------------------------+---------+-----+------+------+--------------+-------------+--------------+----------------+
| StopWordFilterNormalizerBench | apply   | 0   | 1    | 0    | 268,999,040b | 1,285.937ms | +0.74σ       | +0.97%         |
| StopWordFilterNormalizerBench | apply   | 0   | 1    | 1    | 268,999,040b | 1,250.089ms | -1.41σ       | -1.85%         |
| StopWordFilterNormalizerBench | apply   | 0   | 1    | 2    | 268,999,040b | 1,284.863ms | +0.67σ       | +0.88%         |
+-------------------------------+---------+-----+------+------+--------------+-------------+--------------+----------------+
```