@simplechris 2020-09-23
> Hey @simplechris looooooooooooove the optimization, but I think we are assuming too much about our end user by providing default English stop words. Since we can optimize this Transformer without adding stop words, let's have a separate discussion as to whether providing default or example stop words will be helpful or harmful to the end-user.

Yep, fair points @andrewdalpino. I agree. I've projected my own use-case/context into this PR without considering other use-cases equally. I'll rebase this PR, remove the defaults, but keep the optimization. 🚀