@simplechris 2020-09-23
@andrewdalpino I've rebased this PR. The optimization remains, and the english stopwords are now just an implementation detail of the benchmark script 👍