@andrewdalpino 2020-09-26
Waaaaaaay faster! Nice job @simplechris thanks for the PR!

![Screenshot 2020-09-26 031412](https://user-images.githubusercontent.com/18690561/94336151-71601e00-ffa6-11ea-8700-bfde55be28eb.png)